+++ src/Kalabean.Domain/Repositories/IRequirementRepository.cs
@@ -12,6 +12,6 @@ namespace Kalabean.Domain.Repositories
     {
         Task<Requirement> GetById(long id, bool includeDeleted = false);
         Task<IQueryable<Requirement>> Get(GetRequirementsRequest request, bool includeDeleted = false);
-        Task<long> Count(GetRequirementsRequest request, bool includeDeleted = false);
+        Task<int> Count(GetRequirementsRequest request, bool includeDeleted = false);
     }
 }
